Revert "refactor: migrate breadcrumb"
This reverts commit 083201419daf71b63916e96c4172be54f43105bd.
