--- src/angular/breadcrumb/index.ts
@@ -1,2 +1,0 @@
-export * from './breadcrumb-group/breadcrumb-group';
-export * from './breadcrumb/breadcrumb';
