+++ src/showcase/stories/angular/breadcrumb/breadcrumb-group/breadcrumb-group.stories.ts
@@ -1,4 +1,5 @@
-import { SbbBreadcrumb, SbbBreadcrumbGroup } from '@sbb-esta/lyne-angular/breadcrumb';
+import { SbbBreadcrumb } from '@sbb-esta/lyne-angular/breadcrumb/breadcrumb';
+import { SbbBreadcrumbGroup } from '@sbb-esta/lyne-angular/breadcrumb/breadcrumb-group';
 import type { Args, Meta } from '@storybook/angular';
 import { argsToTemplate, moduleMetadata } from '@storybook/angular';
 import { withActions } from 'storybook/actions/decorator';
